# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Procurement Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenders:
    get:
      operationId: listTenders
      tags: [tenders]
      summary: List and search tenders
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Keyword search
        - name: category
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, awarded, cancelled]
        - name: minValue
          in: query
          schema:
            type: number
        - name: maxValue
          in: query
          schema:
            type: number
        - name: source
          in: query
          schema:
            type: string
            enum: [gem, cppp, state, psu, railway, defence, municipal, private]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of tenders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenderListResponse"

    post:
      operationId: createTender
      tags: [tenders]
      summary: Manually add a tender
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenderRequest"
      responses:
        "201":
          description: Tender created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tender"
  /tenders/{id}:
    get:
      operationId: getTender
      tags: [tenders]
      summary: Get tender details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tender detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenderDetail"
        "404":
          description: Not found

    put:
      operationId: updateTender
      tags: [tenders]
      summary: Update tender
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenderRequest"
      responses:
        "200":
          description: Updated tender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tender"

    delete:
      operationId: deleteTender
      tags: [tenders]
      summary: Delete tender
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /tenders/{id}/track:
    post:
      operationId: trackTender
      tags: [tenders]
      summary: Add tender to tracked bids
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tracked

  # ── ANALYSIS ─────────────────────────────────────────────────────────────────
...
  /proposals:
    get:
      operationId: listProposals
      tags: [proposals]
      summary: List proposals
      parameters:
        - name: bidId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Proposals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"

    post:
      operationId: createProposal
      tags: [proposals]
      summary: Generate or create a proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProposalRequest"
      responses:
        "201":
          description: Proposal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get proposal detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"

    put:
      operationId: updateProposal
      tags: [proposals]
      summary: Update proposal content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProposalRequest"
      responses:
        "200":
          description: Updated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"

  # ── BOQ ───────────────────────────────────────────────────────────────────────
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      summary: List vendors and OEM contacts
      responses:
        "200":
          description: Vendors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vendor"

    post:
      operationId: createVendor
      tags: [vendors]
      summary: Add vendor/OEM contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"
  /vendors/{id}:
    put:
      operationId: updateVendor
      tags: [vendors]
      summary: Update vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVendorRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vendor"

    delete:
      operationId: deleteVendor
      tags: [vendors]
      summary: Delete vendor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Tender:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        title:
          type: string
        referenceNumber:
          type: string
        authority:
          type: string
        category:
          type: string
        state:
          type: string
        estimatedValue:
          type: number
        emdAmount:
          type: number
        status:
          type: string
          enum: [open, closed, awarded, cancelled]
        source:
          type: string
          enum: [gem, cppp, state, psu, railway, defence, municipal, private]
        openingDate:
          type: string
          format: date
        closingDate:
          type: string
          format: date
        preBidDate:
          type: string
          format: date
        riskScore:
          type: string
          enum: [green, yellow, red]
        aiSummary:
          type: string
        isTracked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, title, authority, category, status, source]
    TenderDetail:
      allOf:
        - $ref: "#/components/schemas/Tender"
        - type: object
          properties:
            description:
              type: string
            eligibilityCriteria:
              type: string
            technicalSpecs:
              type: string
            paymentTerms:
              type: string
            penalties:
              type: string
            hiddenClauses:
              type: string
            documents:
              type: array
              items:
                $ref: "#/components/schemas/TenderDocument"
            amendments:
              type: array
              items:
                $ref: "#/components/schemas/Amendment"
    TenderDocument:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        type:
          type: string
      required: [id, name, url]
    TenderListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Tender"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateTenderRequest:
      type: object
      properties:
        title:
          type: string
        referenceNumber:
          type: string
        authority:
          type: string
        category:
          type: string
        state:
          type: string
        estimatedValue:
          type: number
        emdAmount:
          type: number
        source:
          type: string
        openingDate:
          type: string
          format: date
        closingDate:
          type: string
          format: date
        preBidDate:
          type: string
          format: date
        description:
          type: string
        eligibilityCriteria:
          type: string
      required: [title, authority, category, source]
    UpdateTenderRequest:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        riskScore:
          type: string
        aiSummary:
          type: string
...
    Proposal:
      type: object
      properties:
        id:
          type: integer
        bidId:
          type: integer
        version:
          type: integer
        type:
          type: string
          enum: [cover_letter, technical, compliance_matrix, deviation, declaration, full]
        title:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [draft, review, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required: [id, bidId, version, type, title, status]
    CreateProposalRequest:
      type: object
      properties:
        bidId:
          type: integer
        type:
          type: string
          enum: [cover_letter, technical, compliance_matrix, deviation, declaration, full]
        title:
          type: string
        aiGenerate:
          type: boolean
      required: [bidId, type, title]
    UpdateProposalRequest:
      type: object
      properties:
        content:
          type: string
        status:
          type: string
          enum: [draft, review, approved, rejected]
        title:
          type: string
...
    Amendment:
      type: object
      properties:
        id:
          type: integer
        tenderId:
          type: integer
        version:
          type: integer
        type:
          type: string
          enum: [deadline_change, spec_change, boq_change, general]
        summary:
          type: string
        changes:
          type: string
        publishedAt:
          type: string
          format: date-time
        detectedAt:
          type: string
          format: date-time
      required: [id, tenderId, version, type, summary]
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        type:
          type: string
          enum: [vendor, oem, subcontractor]
        categories:
          type: array
          items:
            type: string
        oemProducts:
          type: array
          items:
            type: string
        rating:
          type: number
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, companyName, type]
    CreateVendorRequest:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        email:
          type: string
        phone:
          type: string
        type:
          type: string
          enum: [vendor, oem, subcontractor]
        categories:
          type: array
          items:
            type: string
        oemProducts:
          type: array
          items:
            type: string
        rating:
          type: number
        notes:
          type: string
      required: [companyName, type]